+++ Assets/TinaX/Core/Extend/UnityEngine/GameObject/GameObjectExt.cs
@@ -13,6 +13,16 @@ namespace TinaX
         }
 
         /// <summary>
+        /// Destroy self, with delay time
+        /// </summary>
+        /// <param name="selfObj"></param>
+        /// <param name="t">delay time </param>
+        public static void DestroySelf(this GameObject selfObj,float t)
+        {
+            GameObject.Destroy(selfObj,t);
+        }
+
+        /// <summary>
         /// 显示
         /// </summary>
         /// <param name="selfObj"></param>
@@ -78,6 +88,22 @@ namespace TinaX
         }
 
         /// <summary>
+        /// 获取Component，如果不存在就新建
+        /// </summary>
+        /// <param name="obj"></param>
+        /// <param name="type">Component type</param>
+        /// <returns></returns>
+        public static Component GetComponentOrAdd(this GameObject obj,System.Type type)
+        {
+            var c = obj.GetComponent(type);
+            if(c == null)
+            {
+                c = obj.AddComponent(type);
+            }
+            return c;
+        }
+
+        /// <summary>
         /// [链式]移除Component, 如果它存在的话
         /// </summary>
         /// <typeparam name="T"></typeparam>
@@ -85,6 +111,38 @@ namespace TinaX
         public static GameObject RemoveComponentIfExists<T>(this GameObject obj) where T : Component
         {
             var t = obj.GetComponent<T>();
+
+            if (t != null)
+            {
+                Object.Destroy(t);
+            }
+            return obj;
+        }
+
+        /// <summary>
+        /// [链式]移除Component, 如果它存在的话
+        /// </summary>
+        /// <param name="obj"></param>
+        /// <param name="type">Component type</param>
+        public static GameObject RemoveComponentIfExists(this GameObject obj, System.Type type)
+        {
+            var t = obj.GetComponent(type);
+
+            if (t != null)
+            {
+                Object.Destroy(t);
+            }
+            return obj;
+        }
+
+        /// <summary>
+        /// [链式]移除Component, 如果它存在的话
+        /// </summary>
+        /// <param name="obj"></param>
+        /// <param name="type">Component type</param>
+        public static GameObject RemoveComponentIfExists(this GameObject obj, string type)
+        {
+            var t = obj.GetComponent(type);
 
             if (t != null)
             {
@@ -108,6 +166,19 @@ namespace TinaX
             }
             return obj;
         }
+
+
+        public static GameObject RemoveComponentsIfExists<T>(this GameObject obj,System.Type type)
+        {
+            var t = obj.GetComponents(type);
+
+            for (var i = 0; i < t.Length; i++)
+            {
+                Object.Destroy(t[i]);
+            }
+            return obj;
+        }
+
 
 
         /// <summary>
@@ -146,13 +217,12 @@ namespace TinaX
         /// 获取或创建GameObject
         /// </summary>
         /// <param name="selfGo"></param>
-        public static GameObject FindOrCreateGo(this GameObject selfGo,string GameObjectName)
+        public static GameObject FindOrCreateGameObject(this GameObject selfGo, string GameObjectName)
         {
             var trans = selfGo.transform.Find(GameObjectName);
             if (trans == null)
             {
-                var go = new GameObject(GameObjectName);
-                go.transform.SetParent(selfGo.transform);
+                var go = new GameObject(GameObjectName).SetParent(selfGo);
                 return go;
             }
             else
@@ -160,6 +230,48 @@ namespace TinaX
                 return trans.gameObject;
             }
         }
+
+        /// <summary>
+        /// 获取或创建GameObject
+        /// </summary>
+        /// <param name="selfGo"></param>
+        public static GameObject FindOrCreateGo(this GameObject selfGo,string GameObjectName)
+        {
+            return selfGo.FindOrCreateGameObject(GameObjectName);
+        }
+
+        /// <summary>
+        /// 获取或创建GameObject
+        /// </summary>
+        /// <param name="selfGo"></param>
+        public static GameObject FindOrCreateGameObject(this GameObject selfGo, string GameObjectName, params System.Type[] Components)
+        {
+            var trans = selfGo.transform.Find(GameObjectName);
+            if (trans == null)
+            {
+                var go = new GameObject(GameObjectName, Components).SetParent(selfGo);
+                return go;
+            }
+            else
+            {
+                return trans.gameObject;
+            }
+        }
+
+        public static GameObject CreateGameObject(this GameObject selfGo, string GameObjectName)
+        {
+            var go = new GameObject(GameObjectName).SetParent(selfGo);
+            return go;
+        }
+
+        public static GameObject CreateGameObject(this GameObject selfGo,string GameObjectName,params System.Type[] Components)
+        {
+            var go = new GameObject(GameObjectName, Components).SetParent(selfGo);
+            return go;
+        }
+
+
+
 
         /// <summary>
         /// 设置父级GameObject
@@ -171,6 +283,8 @@ namespace TinaX
             return selfGo;
         }
 
+        
+
 
         /// <summary>
         /// 设置世界坐标
@@ -178,7 +292,10 @@ namespace TinaX
         /// <param name="selfGo"></param>
         public static GameObject SetPosition(this GameObject selfGo, Vector3 position)
         {
-            selfGo.transform.position = position;
+            if(selfGo.transform != null)
+            {
+                selfGo.transform.position = position;
+            }
             return selfGo;
         }
 
@@ -189,9 +306,51 @@ namespace TinaX
         /// <param name="selfGo"></param>
         public static GameObject SetLocalPosition(this GameObject selfGo, Vector3 localPosition)
         {
-            selfGo.transform.localPosition = localPosition;
+            if (selfGo.transform != null)
+            {
+                selfGo.transform.localPosition = localPosition;
+
+            }
             return selfGo;
         }
+
+        public static GameObject SetLocalScale(this GameObject selfGo,Vector3 scaleValue)
+        {
+            if(selfGo.transform != null)
+            {
+                selfGo.transform.localScale = scaleValue;
+            }
+            return selfGo;
+        }
+
+        public static GameObject SetRotation(this GameObject selfGo, Quaternion value)
+        {
+            if (selfGo.transform != null)
+            {
+                selfGo.transform.rotation = value;
+            }
+            return selfGo;
+        }
+
+        public static GameObject SetLocalRotation(this GameObject selfGo, Quaternion value)
+        {
+            if (selfGo.transform != null)
+            {
+                selfGo.transform.localRotation = value;
+            }
+            return selfGo;
+        }
+
+        public static GameObject SetEulerAngles(this GameObject selfGo, Vector3 value)
+        {
+            if (selfGo.transform != null)
+            {
+                selfGo.transform.eulerAngles = value;
+            }
+            return selfGo;
+        }
+
+        
 
 
         /// <summary>
@@ -202,6 +361,28 @@ namespace TinaX
         public static bool HasComponent<T>(this GameObject gameObject) where T : Component
         {
             return gameObject.GetComponent<T>() != null;
+        }
+
+        /// <summary>
+        /// 检查组件是否存在
+        /// </summary>
+        /// <param name="gameObject">Game object</param>
+        /// <param name="type">组件类型</param>
+        /// <returns>True when component is attached.</returns>
+        public static bool HasComponent(this GameObject gameObject,string type)
+        {
+            return gameObject.GetComponent(type) != null;
+        }
+
+        /// <summary>
+        /// 检查组件是否存在
+        /// </summary>
+        /// <param name="gameObject">Game object</param>
+        /// <param name="type">组件类型</param>
+        /// <returns>True when component is attached.</returns>
+        public static bool HasComponent(this GameObject gameObject, System.Type type)
+        {
+            return gameObject.GetComponent(type) != null;
         }
 
         /// <summary>